Added test for slim_string, fixed a couple of bugs in copy and substr.
git-svn-id: https://boost-log.svn.sourceforge.net/svnroot/boost-log/trunk/boost-log@224 ab6a8de6-4e00-4c3d-a860-8899c17d353f

+++ libs/log/test/run/util_dynamic_type_dispatcher.cpp
@@ -130,7 +130,7 @@ BOOST_AUTO_TEST_CASE(type_dispatch)
     BOOST_CHECK(val2.dispatch(disp));
 
     // This one is not supported
-    my_value< float > val3(-4.3);
+    my_value< float > val3(static_cast< float >(-4.3));
     vis.set_expected();
     BOOST_CHECK(!val3.dispatch(disp));
 }
